+++ src/wai/annotations/adams/__init__.py
@@ -1,7 +1,6 @@
 from ._format import ADAMSExternalFormat
 from ._prefixes import (
     find_all_prefixes,
-    add_prefix_as_metadata,
     divide_by_prefix,
     remove_prefix_metadata
 )